+++ lang/fr/chunky.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+return [
+    'http' => [
+        'upload_not_found' => 'Téléversement introuvable.',
+        'upload_finalized' => 'Téléversement introuvable ou déjà terminé.',
+        'batch_not_found' => 'Lot introuvable.',
+        'busy' => 'Téléversement temporairement occupé, veuillez réessayer.',
+    ],
+];