+++ packages/core/src/index.ts
@@ -37,7 +37,6 @@ export type {
     BatchResult,
     BatchUploaderState,
     BatchUploaderEventMap,
-    EventCallback,
     JsonPrimitive,
     JsonValue,
     JsonObject,