+++ routes/api.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 use Illuminate\Support\Facades\Route;
 use NETipar\Chunky\Http\Controllers\BatchStatusController;
 use NETipar\Chunky\Http\Controllers\CancelBatchController;